Simplify calculation of runtime costs.
